+++ src/rules/selector-no-type/__tests__/index.js
@@ -27,6 +27,7 @@ testRule(true, tr => {
   tr.ok(".foo { &:nth-child(3n + 1) {} }")
   tr.ok("@keyframes spin { 0% {} }")
   tr.ok("@keyframes spin { to {} from {} }")
+  tr.ok(":root { --custom-property-set: {} }")
 
   tr.notOk("foo {}", {
     message: messages.rejected,